+++ src/domain/props/GradientLegendContainerItem.ts
