+++ src/domain/props/PointLegendContainerItem.ts
@@ -1,6 +1,6 @@
 import PointLegendRecord from './PointLegendRecord';
 
-export default interface PointLegendItem {
+export default interface PointLegendContainerItem {
   title: string;
   records: PointLegendRecord[];
   tooltipInfo?: string;
